+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Fixed
+
 * Update dependencies
+
+### Changed
+
+* (breaking change) dist: Add cargo build option support
 
 ## [0.6.1] - 2022-12-03